% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hexjson.R
\name{convert_and_save_hexjson}
\alias{convert_and_save_hexjson}
\title{Convert tabular data to hexjson and save to a file}
\usage{
convert_and_save_hexjson(data, filename, layout = "odd-r")
}
\arguments{
\item{data}{A dataframe of labels and data to store in each hex.}

\item{filename}{The name of an output file for the hexjson.}

\item{layout}{The coordinate layout of the hexsjon. Must be one of:
odd-r, even-r, odd-q, even-q.}
}
\description{
Converts a dataframe of codes, names, hexjson coordinates, and other data to
a hexjson string and then saves it to a file. The values in the first column
are used as the key for each hex in the hexjson and therefore must be unique.
The values in columns \code{q} and \code{r} are the hexjson coordinates for
columns and rows respectively. The values in any other columns are stored as
properties of each hex.
}
