% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hexjson.R
\name{convert_hexjson}
\alias{convert_hexjson}
\title{Convert tabular data to hexjson}
\usage{
convert_hexjson(data, layout = "odd-r")
}
\arguments{
\item{data}{A dataframe of labels and data to store in each hex.}

\item{layout}{The coordinate layout of the hexsjon. Must be one of:
odd-r, even-r, odd-q, even-q.}
}
\value{
A hexjson string
}
\description{
Converts a dataframe of codes, names, hexjson coordinates, and other data to
a hexjson string. The values in the first column are used as the key for
each hex in the hexjson and therefore must be unique. The values in columns
\code{q} and \code{r} are the hexjson coordinates for columns and rows
respectively. The values in any other columns are stored as properties of
each hex.
}
