% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hexjson.R
\name{convert_hexjson_from_csv}
\alias{convert_hexjson_from_csv}
\title{Convert a csv of data to hexjson and save to a file}
\usage{
convert_hexjson_from_csv(csv_file, hexjson_file, layout = "odd-r")
}
\arguments{
\item{csv_file}{The name of a csv file of data to store in each hex.}

\item{hexjson_file}{The name of an output file for the hexjson.}

\item{layout}{The coordinate layout of the hexsjon. Must be one of:
odd-r, even-r, odd-q, even-q.}
}
\description{
Converts a csv of codes, names, hexjson coordinates, and other_data to a
hexjson string and then saves it to a file. The values in the first column
are used as the key for each hex in the hexjson and therefore must be
unique. The values in columns \code{q} and \code{r} are the hexjson
coordinates for columns and rows respectively. The values in any other
columns are stored as properties of each hex.
}
