% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hexjson.R
\name{create_hexjson_from_csv}
\alias{create_hexjson_from_csv}
\title{Convert a csv of data to hexjson, add unique coordinates, and save to a file}
\usage{
create_hexjson_from_csv(csv_file, hexjson_file, layout = "odd-r")
}
\arguments{
\item{csv_file}{The name of a csv file of data to store in each hex.}

\item{hexjson_file}{The name of an output file for the hexjson.}

\item{layout}{The coordinate layout of the hexsjon. Must be one of:
odd-r, even-r, odd-q, even-q.}
}
\description{
Converts a csv of codes, names and other data to a hexjson string, adding
unique column and row coordinates for each hex, then saves it to a file.
The values in the first column are used as the key for each hex in the
hexjson and therefore must be unique. This function can be used to produce
initial hexjson output which can then be edited in a hexjson editor.
}
